[add]:添加校准角度功能

+++ app/src/main/java/cn/com/medicalmeasurementassistant/listener/DeviceInfoListener.java
@@ -17,7 +17,7 @@ public interface DeviceInfoListener {
 
     void replyStopDataCollect(List<Integer> data);
 
-    void replyCapacitance(float capacitance);
+    void replyCapacitance(double capacitance);
 
     void replyDeviceStopped();
 }
